Unify MOTPE and TPE
