+++ optuna/samplers/_tpe/multi_objective_sampler.py
@@ -1,26 +1,10 @@
-import math
-from typing import Any
 from typing import Callable
-from typing import cast
-from typing import Dict
-from typing import List
-from typing import NamedTuple
 from typing import Optional
-from typing import Sequence
-from typing import Tuple
 
 import numpy as np
-import scipy.special
-from scipy.stats import truncnorm
 
-import optuna
-from optuna import distributions
-from optuna._experimental import experimental
-from optuna.distributions import BaseDistribution
-from optuna.samplers._base import BaseSampler
-from optuna.samplers._random import RandomSampler
-from optuna.samplers._search_space import IntersectionSearchSpace
-from optuna.study import StudyDirection
+from optuna._deprecated import deprecated
+from optuna.samplers._tpe.sampler import TPESampler
 
 
 EPS = 1e-12
@@ -34,129 +18,8 @@ def _default_weights_above(x: int) -> np.ndarray:
     return np.ones(x)
 
 
-class _ParzenEstimatorParameters(
-    NamedTuple(
-        "_ParzenEstimatorParameters",
-        [
-            ("consider_prior", bool),
-            ("prior_weight", Optional[float]),
-            ("consider_magic_clip", bool),
-            ("consider_endpoints", bool),
-            ("weights", Callable[[int], np.ndarray]),
-        ],
-    )
-):
-    pass
-
-
-class _ParzenEstimator(object):
-    def __init__(
-        self, mus: np.ndarray, low: float, high: float, parameters: _ParzenEstimatorParameters
-    ) -> None:
-
-        self.weights, self.mus, self.sigmas = _ParzenEstimator._calculate(
-            mus,
-            low,
-            high,
-            parameters.consider_prior,
-            parameters.prior_weight,
-            parameters.consider_magic_clip,
-            parameters.consider_endpoints,
-            parameters.weights,
-        )
-
-    @classmethod
-    def _calculate(
-        cls,
-        mus: np.ndarray,
-        low: float,
-        high: float,
-        consider_prior: bool,
-        prior_weight: Optional[float],
-        consider_magic_clip: bool,
-        consider_endpoints: bool,
-        weights_func: Callable[[int], np.ndarray],
-    ) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
-        """Calculates the weights, mus and sigma for the Parzen estimator.
-
-        Note: When the number of observations is zero, the Parzen estimator ignores the
-        `consider_prior` flag and utilizes a prior. Validation of this approach is future work.
-        """
-
-        mus = np.asarray(mus)
-        sigma = np.asarray([], dtype=float)
-        prior_pos = 0
-
-        # Parzen estimator construction requires at least one observation or a prior.
-        if mus.size == 0:
-            consider_prior = True
-
-        if consider_prior:
-            prior_mu = 0.5 * (low + high)
-            prior_sigma = 1.0 * (high - low)
-            if mus.size == 0:
-                low_sorted_mus_high = np.zeros(3)
-                sorted_mus = low_sorted_mus_high[1:-1]
-                sorted_mus[0] = prior_mu
-                sigma = np.asarray([prior_sigma])
-                prior_pos = 0
-                order = np.array([]).astype(int)
-            else:  # When mus.size is greater than 0.
-                # We decide the place of the  prior.
-                order = np.argsort(mus)
-                ordered_mus = mus[order]
-                prior_pos = int(np.searchsorted(ordered_mus, prior_mu))
-                # We decide the mus.
-                low_sorted_mus_high = np.zeros(len(mus) + 3)
-                sorted_mus = low_sorted_mus_high[1:-1]
-                sorted_mus[:prior_pos] = ordered_mus[:prior_pos]
-                sorted_mus[prior_pos] = prior_mu
-                sorted_mus[prior_pos + 1 :] = ordered_mus[prior_pos:]
-        else:
-            order = np.argsort(mus)
-            # We decide the mus.
-            low_sorted_mus_high = np.zeros(len(mus) + 2)
-            sorted_mus = low_sorted_mus_high[1:-1]
-            sorted_mus[:] = mus[order]
-
-        # We decide the sigma.
-        if mus.size > 0:
-            low_sorted_mus_high[-1] = high
-            low_sorted_mus_high[0] = low
-            sigma = np.maximum(
-                low_sorted_mus_high[1:-1] - low_sorted_mus_high[0:-2],
-                low_sorted_mus_high[2:] - low_sorted_mus_high[1:-1],
-            )
-            if not consider_endpoints and low_sorted_mus_high.size > 2:
-                sigma[0] = low_sorted_mus_high[2] - low_sorted_mus_high[1]
-                sigma[-1] = low_sorted_mus_high[-2] - low_sorted_mus_high[-3]
-
-        # We decide the weights.
-        unsorted_weights = weights_func(mus.size)
-        if consider_prior:
-            sorted_weights = np.zeros_like(sorted_mus)
-            sorted_weights[:prior_pos] = unsorted_weights[order[:prior_pos]]
-            sorted_weights[prior_pos] = prior_weight
-            sorted_weights[prior_pos + 1 :] = unsorted_weights[order[prior_pos:]]
-        else:
-            sorted_weights = unsorted_weights[order]
-        sorted_weights /= sorted_weights.sum()
-
-        # We adjust the range of the 'sigma' according to the 'consider_magic_clip' flag.
-        maxsigma = 1.0 * (high - low)
-        if consider_magic_clip:
-            minsigma = 1.0 * (high - low) / min(100.0, (1.0 + len(sorted_mus)))
-        else:
-            minsigma = EPS
-        sigma = np.asarray(np.clip(sigma, minsigma, maxsigma))
-        if consider_prior:
-            sigma[prior_pos] = prior_sigma
-
-        return sorted_weights, sorted_mus, sigma
-
-
-@experimental("2.4.0")
-class MOTPESampler(BaseSampler):
+@deprecated("2.8.0")
+class MOTPESampler(TPESampler):
     """Multi-objective sampler using the MOTPE algorithm.
 
     This sampler is a multiobjective version of :class:`~optuna.samplers.TPESampler`.
@@ -249,671 +112,17 @@ class MOTPESampler(BaseSampler):
         seed: Optional[int] = None,
     ) -> None:
 
-        self._parzen_estimator_parameters = _ParzenEstimatorParameters(
-            consider_prior, prior_weight, consider_magic_clip, consider_endpoints, weights_above
-        )
-        self._prior_weight = prior_weight
-        self._n_startup_trials = n_startup_trials
-        self._n_ehvi_candidates = n_ehvi_candidates
-        self._gamma = gamma
-        self._weights = weights_above
-
-        self._warn_independent_sampling = True
-        self._rng = np.random.RandomState(seed)
-        self._mo_random_sampler = RandomSampler(seed=seed)
-
-        self._search_space = IntersectionSearchSpace(include_pruned=True)
-        self._split_cache: Dict[int, Any] = {}
-        self._weights_below: Dict[int, Any] = {}
-
-    def reseed_rng(self) -> None:
-        self._rng = np.random.RandomState()
-        self._mo_random_sampler.reseed_rng()
-
-    def infer_relative_search_space(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-    ) -> Dict[str, BaseDistribution]:
-        return {}
-
-    def sample_relative(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        search_space: Dict[str, BaseDistribution],
-    ) -> Dict[str, Any]:
-        return {}
-
-    def sample_independent(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        param_name: str,
-        param_distribution: BaseDistribution,
-    ) -> Any:
-        if len(study.directions) < 2:
-            raise ValueError(
-                "Number of objectives must be >= 2. "
-                "Please use optuna.samplers.self for single-objective optimization."
-            )
-
-        values, scores = _get_observation_pairs(study, param_name)
-        n = len(values)
-        if n < self._n_startup_trials:
-            return self._mo_random_sampler.sample_independent(
-                study, trial, param_name, param_distribution
-            )
-        below_param_values, above_param_values = self._split_mo_observation_pairs(
-            study, trial, values, scores
+        super().__init__(
+            consider_prior=consider_prior,
+            prior_weight=prior_weight,
+            consider_magic_clip=consider_magic_clip,
+            consider_endpoints=consider_endpoints,
+            n_startup_trials=n_startup_trials,
+            n_ei_candidates=n_ehvi_candidates,
+            gamma=gamma,
+            weights=weights_above,
+            seed=seed,
         )
 
-        if isinstance(param_distribution, distributions.UniformDistribution):
-            return self._sample_mo_uniform(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-        elif isinstance(param_distribution, distributions.LogUniformDistribution):
-            return self._sample_mo_loguniform(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-        elif isinstance(param_distribution, distributions.DiscreteUniformDistribution):
-            return self._sample_mo_discrete_uniform(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-        elif isinstance(param_distribution, distributions.IntUniformDistribution):
-            return self._sample_mo_int(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-        elif isinstance(param_distribution, distributions.IntLogUniformDistribution):
-            return self._sample_mo_int_loguniform(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-        elif isinstance(param_distribution, distributions.CategoricalDistribution):
-            index = self._sample_mo_categorical_index(
-                study, trial, param_distribution, below_param_values, above_param_values
-            )
-            return param_distribution.choices[index]
-        else:
-            distribution_list = [
-                distributions.UniformDistribution.__name__,
-                distributions.LogUniformDistribution.__name__,
-                distributions.DiscreteUniformDistribution.__name__,
-                distributions.IntUniformDistribution.__name__,
-                distributions.IntLogUniformDistribution.__name__,
-                distributions.CategoricalDistribution.__name__,
-            ]
-            raise NotImplementedError(
-                "The distribution {} is not implemented. "
-                "The parameter distribution should be one of the {}".format(
-                    param_distribution, distribution_list
-                )
-            )
 
-    def _split_mo_observation_pairs(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        config_vals: List[Optional[float]],
-        loss_vals: List[List[float]],
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        """Split observations into observations for l(x) and g(x) with the ratio of gamma:1-gamma.
 
-        Weights for l(x) are also calculated in this method.
-
-        This splitting strategy consists of the following two steps:
-            1. Nondonation rank-based selection
-            2. Hypervolume subset selection problem (HSSP)-based selection
-
-        Please refer to the `original paper <https://dl.acm.org/doi/abs/10.1145/3377930.3389817>`_
-        for more details.
-        """
-        cvals = np.asarray(config_vals)
-        lvals = np.asarray(loss_vals)
-
-        # Solving HSSP for variables number of times is a waste of time.
-        # We cache the result of splitting.
-        is_cached = trial._trial_id in self._split_cache
-        if is_cached:
-            split_cache = self._split_cache[trial._trial_id]
-            indices_below = np.asarray(split_cache["indices_below"])
-            weights_below = np.asarray(split_cache["weights_below"])
-            indices_above = np.asarray(split_cache["indices_above"])
-        else:
-            nondomination_ranks = _calculate_nondomination_rank(lvals)
-            n_below = self._gamma(len(lvals))
-            assert 0 <= n_below <= len(lvals)
-
-            indices = np.array(range(len(lvals)))
-            indices_below = np.empty(n_below, dtype=int)
-
-            # Nondomination rank-based selection
-            i = 0
-            last_idx = 0
-            while last_idx + sum(nondomination_ranks == i) <= n_below:
-                length = indices[nondomination_ranks == i].shape[0]
-                indices_below[last_idx : last_idx + length] = indices[nondomination_ranks == i]
-                last_idx += length
-                i += 1
-
-            # Hypervolume subset selection problem (HSSP)-based selection
-            subset_size = n_below - last_idx
-            if subset_size > 0:
-                rank_i_lvals = lvals[nondomination_ranks == i]
-                rank_i_indices = indices[nondomination_ranks == i]
-                worst_point = np.max(rank_i_lvals, axis=0)
-                reference_point = np.maximum(1.1 * worst_point, 0.9 * worst_point)
-                reference_point[reference_point == 0] = EPS
-                selected_indices = self._solve_hssp(
-                    rank_i_lvals, rank_i_indices, subset_size, reference_point
-                )
-                indices_below[last_idx:] = selected_indices
-
-            indices_above = np.setdiff1d(indices, indices_below)
-
-            attrs = {
-                "indices_below": indices_below.tolist(),
-                "indices_above": indices_above.tolist(),
-            }
-            weights_below = self._calculate_weights_below(lvals, indices_below)
-            attrs["weights_below"] = weights_below.tolist()
-            self._split_cache[trial._trial_id] = attrs
-
-        below = cvals[indices_below]
-        self._weights_below[trial._trial_id] = np.asarray(
-            [w for w, v in zip(weights_below, below) if v is not None], dtype=float
-        )
-        below = np.asarray([v for v in below if v is not None], dtype=float)
-        above = cvals[indices_above]
-        above = np.asarray([v for v in above if v is not None], dtype=float)
-        return below, above
-
-    def _sample_mo_uniform(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.UniformDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> float:
-        low = distribution.low
-        high = distribution.high
-        return self._sample_mo_numerical(study, trial, low, high, below, above)
-
-    def _sample_mo_loguniform(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.LogUniformDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> float:
-        low = distribution.low
-        high = distribution.high
-        return self._sample_mo_numerical(study, trial, low, high, below, above, is_log=True)
-
-    def _sample_mo_discrete_uniform(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.DiscreteUniformDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> float:
-        q = distribution.q
-        r = distribution.high - distribution.low
-        # [low, high] is shifted to [0, r] to align sampled values at regular intervals.
-        low = 0 - 0.5 * q
-        high = r + 0.5 * q
-
-        # Shift below and above to [0, r]
-        above -= distribution.low
-        below -= distribution.low
-
-        best_sample = (
-            self._sample_mo_numerical(study, trial, low, high, below, above, q=q)
-            + distribution.low
-        )
-        return min(max(best_sample, distribution.low), distribution.high)
-
-    def _sample_mo_int(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.IntUniformDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> int:
-        d = distributions.DiscreteUniformDistribution(
-            low=distribution.low, high=distribution.high, q=distribution.step
-        )
-        return int(self._sample_mo_discrete_uniform(study, trial, d, below, above))
-
-    def _sample_mo_int_loguniform(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.IntLogUniformDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> int:
-        low = distribution.low - 0.5
-        high = distribution.high + 0.5
-
-        sample = self._sample_mo_numerical(study, trial, low, high, below, above, is_log=True)
-        best_sample = np.round(sample)
-
-        return int(min(max(best_sample, distribution.low), distribution.high))
-
-    def _sample_mo_numerical(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        low: float,
-        high: float,
-        below: np.ndarray,
-        above: np.ndarray,
-        q: Optional[float] = None,
-        is_log: bool = False,
-    ) -> float:
-        if is_log:
-            low = np.log(low)
-            high = np.log(high)
-            below = np.log(below)
-            above = np.log(above)
-
-        size = (self._n_ehvi_candidates,)
-
-        weights_below: Callable[[int], np.ndarray]
-        weights_below = lambda _: self._weights_below[trial._trial_id]  # NOQA
-
-        parzen_estimator_parameters_below = _ParzenEstimatorParameters(
-            self._parzen_estimator_parameters.consider_prior,
-            self._parzen_estimator_parameters.prior_weight,
-            self._parzen_estimator_parameters.consider_magic_clip,
-            self._parzen_estimator_parameters.consider_endpoints,
-            weights_below,
-        )
-        parzen_estimator_below = _ParzenEstimator(
-            mus=below, low=low, high=high, parameters=parzen_estimator_parameters_below
-        )
-        samples_below = self._sample_from_gmm(
-            parzen_estimator=parzen_estimator_below,
-            low=low,
-            high=high,
-            q=q,
-            size=size,
-        )
-        log_likelihoods_below = self._gmm_log_pdf(
-            samples=samples_below,
-            parzen_estimator=parzen_estimator_below,
-            low=low,
-            high=high,
-            q=q,
-        )
-
-        weights_above = self._weights
-        parzen_estimator_parameters_above = _ParzenEstimatorParameters(
-            self._parzen_estimator_parameters.consider_prior,
-            self._parzen_estimator_parameters.prior_weight,
-            self._parzen_estimator_parameters.consider_magic_clip,
-            self._parzen_estimator_parameters.consider_endpoints,
-            weights_above,
-        )
-        parzen_estimator_above = _ParzenEstimator(
-            mus=above, low=low, high=high, parameters=parzen_estimator_parameters_above
-        )
-        log_likelihoods_above = self._gmm_log_pdf(
-            samples=samples_below,
-            parzen_estimator=parzen_estimator_above,
-            low=low,
-            high=high,
-            q=q,
-        )
-
-        ret = float(
-            self._compare(
-                samples=samples_below, log_l=log_likelihoods_below, log_g=log_likelihoods_above
-            )[0]
-        )
-        return math.exp(ret) if is_log else ret
-
-    def _sample_mo_categorical_index(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        distribution: distributions.CategoricalDistribution,
-        below: np.ndarray,
-        above: np.ndarray,
-    ) -> int:
-        choices = distribution.choices
-        below = below.astype(int)
-        above = above.astype(int)
-        upper = len(choices)
-        size = (self._n_ehvi_candidates,)
-
-        weights_below = self._weights_below[trial._trial_id]
-        counts_below = np.bincount(below, minlength=upper, weights=weights_below)
-        weighted_below = counts_below + self._prior_weight
-        weighted_below /= weighted_below.sum()
-        samples_below = self._sample_from_categorical_dist(weighted_below, size)
-        log_likelihoods_below = self._categorical_log_pdf(samples_below, weighted_below)
-
-        weights_above = self._weights(len(above))
-        counts_above = np.bincount(above, minlength=upper, weights=weights_above)
-        weighted_above = counts_above + self._prior_weight
-        weighted_above /= weighted_above.sum()
-        log_likelihoods_above = self._categorical_log_pdf(samples_below, weighted_above)
-
-        return int(
-            self._compare(
-                samples=samples_below, log_l=log_likelihoods_below, log_g=log_likelihoods_above
-            )[0]
-        )
-
-    @staticmethod
-    def _compute_hypervolume(solution_set: np.ndarray, reference_point: np.ndarray) -> float:
-        return optuna.multi_objective._hypervolume.WFG().compute(solution_set, reference_point)
-
-    def _solve_hssp(
-        self,
-        rank_i_loss_vals: np.ndarray,
-        rank_i_indices: np.ndarray,
-        subset_size: int,
-        reference_point: np.ndarray,
-    ) -> np.ndarray:
-        """Solve a hypervolume subset selection problem (HSSP) via a greedy algorithm.
-
-        This method is a 1-1/e approximation algorithm to solve HSSP.
-
-        For further information about algorithms to solve HSSP, please refer to the following
-        paper:
-
-        - `Greedy Hypervolume Subset Selection in Low Dimensions
-           <https://ieeexplore.ieee.org/document/7570501>`_
-        """
-        selected_vecs = []  # type: List[np.ndarray]
-        selected_indices = []  # type: List[int]
-        contributions = [
-            self._compute_hypervolume(np.asarray([v]), reference_point) for v in rank_i_loss_vals
-        ]
-        hv_selected = 0.0
-        while len(selected_indices) < subset_size:
-            max_index = int(np.argmax(contributions))
-            contributions[max_index] = -1  # mark as selected
-            selected_index = rank_i_indices[max_index]
-            selected_vec = rank_i_loss_vals[max_index]
-            for j, v in enumerate(rank_i_loss_vals):
-                if contributions[j] == -1:
-                    continue
-                p = np.max([selected_vec, v], axis=0)
-                contributions[j] -= (
-                    self._compute_hypervolume(np.asarray(selected_vecs + [p]), reference_point)
-                    - hv_selected
-                )
-            selected_vecs += [selected_vec]
-            selected_indices += [selected_index]
-            hv_selected = self._compute_hypervolume(np.asarray(selected_vecs), reference_point)
-
-        return np.asarray(selected_indices, dtype=int)
-
-    def _calculate_weights_below(
-        self,
-        lvals: np.ndarray,
-        indices_below: np.ndarray,
-    ) -> np.ndarray:
-        # Calculate weights based on hypervolume contributions.
-        n_below = len(indices_below)
-        if n_below == 0:
-            return np.asarray([])
-        elif n_below == 1:
-            return np.asarray([1.0])
-        else:
-            lvals_below = lvals[indices_below].tolist()
-            worst_point = np.max(lvals_below, axis=0)
-            reference_point = np.maximum(1.1 * worst_point, 0.9 * worst_point)
-            reference_point[reference_point == 0] = EPS
-            hv = self._compute_hypervolume(np.asarray(lvals_below), reference_point)
-            contributions = np.asarray(
-                [
-                    hv
-                    - self._compute_hypervolume(
-                        np.asarray(lvals_below[:i] + lvals_below[i + 1 :]), reference_point
-                    )
-                    for i in range(n_below)
-                ]
-            )
-            contributions += EPS
-            weights_below = np.clip(contributions / np.max(contributions), 0, 1)
-            return weights_below
-
-    def after_trial(
-        self,
-        study: optuna.study.Study,
-        trial: optuna.trial.FrozenTrial,
-        state: optuna.trial.TrialState,
-        values: Optional[Sequence[float]],
-    ) -> None:
-        if trial._trial_id in self._split_cache:
-            del self._split_cache[trial._trial_id]
-        if trial._trial_id in self._weights_below:
-            del self._weights_below[trial._trial_id]
-
-        self._mo_random_sampler.after_trial(study, trial, state, values)
-
-    def _sample_from_gmm(
-        self,
-        parzen_estimator: _ParzenEstimator,
-        low: float,
-        high: float,
-        q: Optional[float] = None,
-        size: Tuple = (),
-    ) -> np.ndarray:
-
-        weights = parzen_estimator.weights
-        mus = parzen_estimator.mus
-        sigmas = parzen_estimator.sigmas
-        weights, mus, sigmas = map(np.asarray, (weights, mus, sigmas))
-
-        if low >= high:
-            raise ValueError(
-                "The 'low' should be lower than the 'high'. "
-                "But (low, high) = ({}, {}).".format(low, high)
-            )
-
-        active = np.argmax(self._rng.multinomial(1, weights, size=size), axis=-1)
-        trunc_low = (low - mus[active]) / sigmas[active]
-        trunc_high = (high - mus[active]) / sigmas[active]
-        samples = np.full((), fill_value=high + 1.0, dtype=np.float64)
-        while (samples >= high).any():
-            samples = np.where(
-                samples < high,
-                samples,
-                truncnorm.rvs(
-                    trunc_low,
-                    trunc_high,
-                    size=size,
-                    loc=mus[active],
-                    scale=sigmas[active],
-                    random_state=self._rng,
-                ),
-            )
-
-        if q is None:
-            return samples
-        else:
-            return np.round(samples / q) * q
-
-    def _gmm_log_pdf(
-        self,
-        samples: np.ndarray,
-        parzen_estimator: _ParzenEstimator,
-        low: float,
-        high: float,
-        q: Optional[float] = None,
-    ) -> np.ndarray:
-
-        weights = parzen_estimator.weights
-        mus = parzen_estimator.mus
-        sigmas = parzen_estimator.sigmas
-        samples, weights, mus, sigmas = map(np.asarray, (samples, weights, mus, sigmas))
-        if samples.size == 0:
-            return np.asarray([], dtype=float)
-        if weights.ndim != 1:
-            raise ValueError(
-                "The 'weights' should be 1-dimension. "
-                "But weights.shape = {}".format(weights.shape)
-            )
-        if mus.ndim != 1:
-            raise ValueError(
-                "The 'mus' should be 1-dimension. But mus.shape = {}".format(mus.shape)
-            )
-        if sigmas.ndim != 1:
-            raise ValueError(
-                "The 'sigmas' should be 1-dimension. But sigmas.shape = {}".format(sigmas.shape)
-            )
-
-        p_accept = np.sum(
-            weights * (self._normal_cdf(high, mus, sigmas) - self._normal_cdf(low, mus, sigmas))
-        )
-
-        if q is None:
-            distance = samples[..., None] - mus
-            mahalanobis = (distance / np.maximum(sigmas, EPS)) ** 2
-            Z = np.sqrt(2 * np.pi) * sigmas
-            coefficient = weights / Z / p_accept
-            return self._logsum_rows(-0.5 * mahalanobis + np.log(coefficient))
-        else:
-            cdf_func = self._normal_cdf
-            upper_bound = np.minimum(samples + q / 2.0, high)
-            lower_bound = np.maximum(samples - q / 2.0, low)
-            probabilities = np.sum(
-                weights[..., None]
-                * (
-                    cdf_func(upper_bound[None], mus[..., None], sigmas[..., None])
-                    - cdf_func(lower_bound[None], mus[..., None], sigmas[..., None])
-                ),
-                axis=0,
-            )
-            return np.log(probabilities + EPS) - np.log(p_accept + EPS)
-
-    def _sample_from_categorical_dist(
-        self, probabilities: np.ndarray, size: Tuple[int]
-    ) -> np.ndarray:
-
-        if size == (0,):
-            return np.asarray([], dtype=float)
-        assert size
-
-        if probabilities.size == 1 and isinstance(probabilities[0], np.ndarray):
-            probabilities = probabilities[0]
-        assert probabilities.ndim == 1
-
-        n_draws = np.prod(size).item()
-        sample = self._rng.multinomial(n=1, pvals=probabilities, size=n_draws)
-        assert sample.shape == size + probabilities.shape
-        return_val = np.dot(sample, np.arange(probabilities.size)).reshape(size)
-        return return_val
-
-    @classmethod
-    def _categorical_log_pdf(cls, sample: np.ndarray, p: np.ndarray) -> np.ndarray:
-
-        if sample.size:
-            return np.log(np.asarray(p)[sample])
-        else:
-            return np.asarray([])
-
-    @classmethod
-    def _compare(cls, samples: np.ndarray, log_l: np.ndarray, log_g: np.ndarray) -> np.ndarray:
-
-        samples, log_l, log_g = map(np.asarray, (samples, log_l, log_g))
-        if samples.size:
-            score = log_l - log_g
-            if samples.size != score.size:
-                raise ValueError(
-                    "The size of the 'samples' and that of the 'score' "
-                    "should be same. "
-                    "But (samples.size, score.size) = ({}, {})".format(samples.size, score.size)
-                )
-
-            best = np.argmax(score)
-            return np.asarray([samples[best]] * samples.size)
-        else:
-            return np.asarray([])
-
-    @classmethod
-    def _logsum_rows(cls, x: np.ndarray) -> np.ndarray:
-
-        x = np.asarray(x)
-        m = x.max(axis=1)
-        return np.log(np.exp(x - m[:, None]).sum(axis=1)) + m
-
-    @classmethod
-    def _normal_cdf(cls, x: float, mu: np.ndarray, sigma: np.ndarray) -> np.ndarray:
-
-        mu, sigma = map(np.asarray, (mu, sigma))
-        denominator = x - mu
-        numerator = np.maximum(np.sqrt(2) * sigma, EPS)
-        z = denominator / numerator
-        return 0.5 * (1 + scipy.special.erf(z))
-
-
-def _calculate_nondomination_rank(loss_vals: np.ndarray) -> np.ndarray:
-    vecs = loss_vals.copy()
-
-    # Normalize values
-    lb = vecs.min(axis=0, keepdims=True)
-    ub = vecs.max(axis=0, keepdims=True)
-    vecs = (vecs - lb) / (ub - lb)
-
-    ranks = np.zeros(len(vecs))
-    num_unranked = len(vecs)
-    rank = 0
-    while num_unranked > 0:
-        extended = np.tile(vecs, (vecs.shape[0], 1, 1))
-        counts = np.sum(
-            np.logical_and(
-                np.all(extended <= np.swapaxes(extended, 0, 1), axis=2),
-                np.any(extended < np.swapaxes(extended, 0, 1), axis=2),
-            ),
-            axis=1,
-        )
-        vecs[counts == 0] = 1.1  # mark as ranked
-        ranks[counts == 0] = rank
-        rank += 1
-        num_unranked -= np.sum(counts == 0)
-    return ranks
-
-
-def _get_observation_pairs(
-    study: optuna.study.Study,
-    param_name: str,
-) -> Tuple[List[Optional[float]], List[List[float]]]:
-    """Get observation pairs from the study.
-
-    This function collects observation pairs from the complete trials of the study.
-    Pruning is currently not supported.
-    The values for trials that don't contain the parameter named ``param_name`` are set to None.
-
-    Objective values are negated if their directions are maximization and all objectives are
-    treated as minimization in the MOTPE algorithm.
-    """
-
-    trials = study.get_trials(deepcopy=False, states=(optuna.trial.TrialState.COMPLETE,))
-    values = []
-    scores = []
-    for trial in trials:
-        param_value = None  # type: Optional[float]
-        if param_name in trial.params:
-            distribution = trial.distributions[param_name]
-            param_value = distribution.to_internal_repr(trial.params[param_name])
-
-        # Convert all objectives to minimization
-        score = [
-            cast(float, v) if d == StudyDirection.MINIMIZE else -cast(float, v)
-            for d, v in zip(study.directions, trial.values)
-        ]
-
-        values.append(param_value)
-        scores.append(score)
-
-    return values, scores
